requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
cssselect2==0.6.0
daphne==3.0.2
decorator==5.1.1
Django==4.0.4
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-tables2==2.4.1
django-webpack-loader==1.5.0
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.11.1
fonttools==4.33.3
frozenlist==1.3.0
h11==0.13.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
msgpack==1.0.3
multidict==6.0.2
mysqlclient==2.1.0
outcome==1.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
pydyf==0.1.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyphen==0.12.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
selenium==4.1.5
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
tinycss2==1.1.1
traitlets==5.2.0
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
Werkzeug==2.1.2
wsproto==1.1.0
yarl==1.7.2
zope.interface==5.4.0
zopfli==0.2.1
